+++ src/components/Inputs/Button.svelte
@@ -14,7 +14,7 @@
 		onclick?: (e: Event) => void;
 		type?: 'submit' | 'button';
 		as?: 'button' | 'a';
-		href?: '';
+		href?: string;
 		iconClass?: string;
 	}
 </script>
@@ -22,7 +22,8 @@
 <script lang="ts">
 	let { type, text = '', as = 'button', color = 'primary', icon: Icon, ...props }: Props = $props();
 
-	const className = 'flex justify-center items-center space-x-2 rounded-xl px-4 py-2 min-h-[40px]';
+	const className =
+		'select-none flex justify-center items-center space-x-2 rounded-xl px-4 py-2 min-h-[40px]';
 	const textClasses = getTextStyles();
 </script>
 
@@ -37,6 +38,11 @@
 	{...props.data}
 	{type}
 >
+	<!--* Custom Icon *-->
+	{#if !!Icon}
+		<Icon class="{textClasses} size-4 {props.iconClass}"></Icon>
+	{/if}
+
 	{#if text !== ''}
 		<p class={textClasses}>
 			{text}
@@ -45,10 +51,5 @@
 
 	{#if props.children}
 		{@render props.children()}
-	{/if}
-
-	<!--* Custom Icon *-->
-	{#if !!Icon}
-		<Icon class="{textClasses} size-4 {props.iconClass}"></Icon>
 	{/if}
 </svelte:element>