+++ src/components/Modals/Modal.svelte
@@ -9,6 +9,7 @@
 		open?: boolean;
 		as?: string;
 		onsubmit?: () => void;
+		fullScreenOnSmallWindow?: boolean;
 	}
 </script>
 
@@ -19,6 +20,7 @@
 		closeWhenClickBackground = true,
 		open = $bindable(false),
 		as = 'form',
+		fullScreenOnSmallWindow = false,
 		...props
 	}: Props = $props();
 </script>
@@ -30,7 +32,7 @@
 		role="form"
 		transition:fade={{ duration: 300, easing: easeOut }}
 		onclick={closeWhenClickBackground ? stopPropagation((e) => (open = false)) : null}
-		class={`group bg-white-glass-10 fixed inset-0 z-20 flex items-center justify-center overflow-y-auto backdrop-blur-lg outline-none ${props.class}`}
+		class={`group bg-white-glass-10 fixed inset-0 z-20 flex justify-center overflow-y-auto backdrop-blur-lg outline-none ${fullScreenOnSmallWindow ? 'sm:items-center' : 'items-center'} ${props.class}`}
 	>
 		<!-- svelte-ignore a11y_no_noninteractive_element_interactions -->
 		<svelte:element
@@ -39,7 +41,9 @@
 			transition:fly={{ duration: 300, y: 30, easing: easeOut }}
 			onsubmit={preventDefault(stopPropagation(() => props.onsubmit && props.onsubmit()))}
 			onclick={stopPropagation(() => {})}
-			class="bg-muted-background shadow-custom dark:shadow-custom-dark rounded-2xl p-5 {props.contentClass}"
+			class="bg-muted-background shadow-custom dark:shadow-custom-dark flex p-5 {fullScreenOnSmallWindow
+				? 'items-center justify-center max-sm:flex-grow sm:rounded-2xl'
+				: 'rounded-2xl'} {props.contentClass}"
 		>
 			{#if props.children}
 				{@render props.children()}