+++ src/consts/env.ts
@@ -1,0 +1,3 @@
+import { PUBLIC_DEBUG } from '$env/static/public';
+
+export const debug = PUBLIC_DEBUG === 'true' ? true : false;