+++ src/routes/+layout.svelte
@@ -5,6 +5,9 @@
 	import { QueryClient, QueryClientProvider } from '@sveltestack/svelte-query';
 	import ToasterInit from '@cp/Toaster/ToasterInit.svelte';
 	import GlobalInit from '@cp/GlobalInit/GlobalInit.svelte';
+	import ThemeSwitcher from '@cp/Theme/ThemeSwitcher.svelte';
+	import DarkModeSwitcher from '@cp/Theme/DarkModeSwitcher.svelte';
+	import LanguageChanger from '@cp/I18n/LanguageChanger.svelte';
 </script>
 
 <script lang="ts">
@@ -21,6 +24,9 @@
 
 <GlobalInit />
 <ToasterInit where="bottom_right" />
+<ThemeSwitcher class="fixed right-0 bottom-0 z-[1000]" />
+<DarkModeSwitcher class="fixed right-16 bottom-0 z-[1000]" />
+<LanguageChanger class="fixed right-32 bottom-0 z-[1000]" />
 <QueryClientProvider client={queryClient}>
 	{@render children()}
 </QueryClientProvider>