+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 
+- 🐛 fix: bug fixes on radiogroup and my debug tools for changing design
 - 🎉 feat: added https in development environment
 - 🎉 feat: DatePicker is not created with Dropdown
 - 🎉 feat: added DatePicker to open like Dropdown component